Remove the `single` module and rename vec.rs to lib.rs. (Breaking Change)
